requirements.txt
flask
finnhub-python